+++ judge-sdk/src/main/java/com/dwarfeng/judge/sdk/bean/dto/FastJsonSectionReport.java
@@ -8,6 +8,7 @@ import com.dwarfeng.subgrade.stack.bean.dto.Dto;
 import java.util.Date;
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 /**
@@ -31,7 +32,9 @@ public class FastJsonSectionReport implements Dto {
                 sectionReport.getSum(),
                 sectionReport.getExpected(),
                 sectionReport.getVariance(),
-                sectionReport.getJudgerReports().stream().map(FastJsonJudgerReport::of).collect(Collectors.toList())
+                Optional.ofNullable(sectionReport.getJudgerReports()).map(
+                        f -> f.stream().map(FastJsonJudgerReport::of).collect(Collectors.toList())
+                ).orElse(null)
         );
     }